+++ DX12/D3D12Research/D3D12/DynamicResourceAllocator.h
@@ -16,7 +16,7 @@ public:
 	~DynamicResourceAllocator();
 
 	DynamicAllocation Allocate(int size);
-	void Free(int fenceValue);
+	void Free(uint64_t fenceValue);
 
 private:
 	ComPtr<ID3D12Resource> m_pBackingResource;
